test: split phase 4 quote/invoice API tests from test_quotes
